# app-service/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - user-rest-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    put:
      tags:
      - user-rest-controller
      operationId: replaceUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
...
  /api/v1/users/{id}/image/:
    post:
      tags:
      - user-rest-controller
      operationId: createUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
...
  /api/v1/products/:
    get:
      tags:
      - product-rest-controller
      operationId: getProducts
      parameters:
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: location
        in: query
        required: false
        schema:
          type: string
      - name: date
        in: query
        required: false
        schema:
          type: string
      - name: minPrice
        in: query
        required: false
        schema:
          type: string
      - name: maxPrice
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelExtendedProductDTO"
    post:
      tags:
      - product-rest-controller
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/users/:
    get:
      tags:
      - user-rest-controller
      operationId: getUsers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelUserDTO"
...
  /api/v1/users/{id}/image/{imageId}:
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
...
components:
  schemas:
    ImageBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    OrderBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        product:
          $ref: "#/components/schemas/ProductDTO"
        totalPrice:
          type: number
          format: double
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
    ProductBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        price:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        category:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        image:
          $ref: "#/components/schemas/ImageBasicDTO"
        isSold:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          type: string
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surnames:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        surnames:
          type: string
        email:
          type: string
        image:
          $ref: "#/components/schemas/ImageBasicDTO"
        location:
          $ref: "#/components/schemas/LocationDTO"
        createdAt:
          type: string
          format: date
        roles:
          type: array
          items:
            type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductBasicDTO"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderBasicDTO"
...
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        imageFile:
          type: object
          properties:
            binaryStream: {}
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelUserDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
    ExtendedProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        image:
          $ref: "#/components/schemas/ImageBasicDTO"
        isSold:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          type: string
        distance:
          type: integer
          format: int32
        interestingCategory:
          type: boolean
        score:
          type: number
          format: double
    PagedModelExtendedProductDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ExtendedProductDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...